Hack in some 8-bit PWM audio. This is not noise corrected, as i don't have any code handy to do that an arbitrary frequency, and i don't feel like rescaling from the OPL2 frequency of 49716. The 8 bit PWM code i had lying around but is a mess.
Bit of a shame, as the sound on RP2040 Doom is so nice. The speaker on the DEF CON 32 Badge
is not great, and seems to distort terribly at moderate volume levels.

This could certainly be user error -- maybe it could use some filtering -- but i'm just
defaulting it to really quite for now - it does add a tiny bit more atmosphere